pyproject.toml
[project]
name = "moviescrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot>=21.10",
]
